change the system of checking access role, and add new module, and new method for client for getting automaticaly all enpoints of server, also set each controller auth guard and role guard

+++ src/roles/roles.service.ts
@@ -1,9 +1,10 @@
 import { forwardRef, HttpException, HttpStatus, Inject, Injectable } from "@nestjs/common";
 import { InjectModel } from "@nestjs/sequelize";
 import { RolesModel } from "./roles.model";
-import { RolesDto } from "./dto/roles.dto";
+import {RolesDto, RolesParamsDto} from "./dto/roles.dto";
 import { SetDoctorsRoles } from "./dto/set-doctors-roles.dto";
 import { DoctorsService } from "../doctors/doctors.service";
+import {UpdateRoleDto} from "./dto/update-role.dto";
 
 @Injectable()
 export class RolesService {
@@ -21,9 +22,18 @@ export class RolesService {
         }
     }
 
-    async getAllRoles(){
+    async getAllRoles(dto?: RolesParamsDto){
         try {
-            return await this.rolesRepository.findAll()
+            if(dto.all === "true"){
+                return await this.rolesRepository.findAll({include: {all: true}});
+            }else if(dto.limit){
+                return await this.rolesRepository.findAll({
+                    limit: dto.limit,
+                    offset: ((dto.page-1) || 0) * dto.limit
+                })
+            }else{
+                return await this.rolesRepository.findAll()
+            }
         }catch (e){
             throw new HttpException({message: e}, HttpStatus.INTERNAL_SERVER_ERROR)
         }
@@ -48,8 +58,27 @@ export class RolesService {
         }
     }
 
+    async updateRole(dto: UpdateRoleDto){
+        try {
+            const role = await this.rolesRepository.findByPk(Number(dto.id))
+            if(role.role === "admin"){
+                throw new HttpException({message: 'that role cant be updated'}, HttpStatus.BAD_REQUEST)
+            }
+
+            return await this.rolesRepository.update( {role: dto.role}, {where: {id: Number(dto.id)},})
+        }catch (e){
+            throw new HttpException({message: e}, HttpStatus.INTERNAL_SERVER_ERROR)
+        }
+    }
+
     async deleteOneRole(id: number){
         try {
+
+            const role = await this.rolesRepository.findByPk(Number(id))
+            if(role.role === "admin"){
+                throw new HttpException({message: 'that role cant be deleted'}, HttpStatus.BAD_REQUEST)
+            }
+
             return await this.rolesRepository.destroy({where: {id: id}})
         }catch (e){
             throw new HttpException({message: e}, HttpStatus.INTERNAL_SERVER_ERROR)
@@ -78,5 +107,18 @@ export class RolesService {
             throw new HttpException({message: e}, HttpStatus.INTERNAL_SERVER_ERROR)
         }
     }
+
+    async getAmountOfRoles(){
+        try {
+            const response = await this.rolesRepository.findAll()
+            if(!response){
+                return 0
+            }
+            return response.length
+        }catch (e){
+            throw new HttpException({message: e}, HttpStatus.INTERNAL_SERVER_ERROR)
+        }
+    }
+
 
 }
